refactor: extract ToolApprovalHandler class from executor.ts
